backend/requirements.txt
Flask
Flask-Cors
python-dotenv
pandas
opencv-python
gunicorn
svgwrite
svgpathtools